Fix neko-vpn WebRTC by removing conflicting EPR setting
Remove NEKO_WEBRTC_EPR (ephemeral port range) which conflicts with
NEKO_WEBRTC_UDPMUX. When using UDPMUX for single-port multiplexing,
EPR should not be set as it causes neko to advertise ICE candidates
for ports 52000-52100, but only 52000 is exposed on the LoadBalancer.

This was causing WebRTC ICE connection failures during the 'checking'
phase.